requirements.txt
diffrax==0.6.2
equinox==0.11.11
iisignature==0.24
jax[cuda12]==0.4.38
jaxtyping==0.2.36
matplotlib==3.10.0
numpy==2.2.2
optax==0.2.4
pyparsing==3.2.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.1
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchdiffeq==0.2.5
tqdm==4.67.1
scienceplots